+++ R/getfd.R
@@ -1,9 +1,21 @@
 #' @title getfd
-#' @description Return normed path to current file directory
+#' @description Return full path to current file directory
+#' @param on.error Expression to use if the current file directory cannot be
+#' determined. In that case, `normalizePath(<on.error>, winslash)` is returned.
+#' Can also be an expression like `stop("message")` to stop execution
+#' (default).
+#' @param winslash Path separator to use for windows
 #' @return Current file directory as string
-#' @examples getfd()
+#' @examples \dontrun{getfd()}
+#' getfd(on.error=getwd())
 #' @export
-getfd <- function() {
+getfd <- function(
+    on.error=stop(
+        "No file sourced. Maybe you're in an interactive shell?",
+        call.=F
+    ),
+    winslash="/"
+) {
     if (interactive() && sys.nframe() > 0 && "ofile" %in% ls(sys.frame(0))) {
         # probably started through `source` from interactive R session
         f <- sys.frame(1)$ofile
@@ -12,12 +24,11 @@ getfd <- function() {
         args <- commandArgs()
         filearg_idx <- grepl("^--file=", args)
         if (any(filearg_idx)) {
-            filearg <- args[][[1]]
+            filearg <- args[filearg_idx][[1]]
             f <- strsplit(filearg, "--file=")[[1]][2]
         } else {
-            stop("No file sourced. Maybe you're in an interactive shell?",
-                 call.=F)
+            return(normalizePath(on.error, winslash=winslash))
         }
     }
-    return(normalizePath(dirname(f), winslash = "/"))
+    return(normalizePath(dirname(f), winslash=winslash))
 }