+++ CHANGELOG.md
@@ -16,6 +16,12 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
   - `Fixed` for bug fixes
   - `Security` in case of vulnerabilities
 
+## [1.3.0] - 2021-12-01
+
+- `Fixed`: function `getfd`. `getfd` now returns the correct file directory also for scripts started through `Rscript`
+- `Added`: optional parameter `on.error` to `getfd`
+- `Added`: optional parameter `winslash` to `getfd`
+
 ## [1.2.0] - 2021-12-01
 
 - `Added`: `cat0`, `catn` and `cat0n` functions